DataFlowDirectory/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
DateTime==5.3
Django==4.2.7
django-rest-framework==0.1.0
djangorestframework==3.14.0
idna==3.6
npm==0.1.1
optional-django==0.1.0
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.1.0
zope.interface==6.1
